+++ tiled/_tests/conftest.py
@@ -13,7 +13,10 @@ from sqlalchemy.ext.asyncio import create_async_engine
 
 from .. import profiles
 from ..catalog import from_uri, in_memory
+from ..client import Context
 from ..client.base import BaseClient
+from ..config import Authentication
+from ..server.app import build_app
 from ..server.settings import get_settings
 from ..utils import ensure_specified_sql_driver
 from .utils import enter_username_password as utils_enter_uname_passwd
@@ -308,6 +311,36 @@ def redis_uri():
         client.flushdb()
     else:
         raise pytest.skip("No TILED_TEST_REDIS configured")
+
+
+@pytest.fixture(scope="function")
+def tiled_websocket_context(tmpdir, redis_uri):
+    """Fixture that provides a Tiled context with websocket support."""
+    tree = from_uri(
+        "sqlite:///:memory:",
+        writable_storage=[
+            f"file://localhost{str(tmpdir / 'data')}",
+            f"duckdb:///{tmpdir / 'data.duckdb'}",
+        ],
+        readable_storage=None,
+        init_if_not_exists=True,
+        # This uses shorter defaults than the production defaults. Nothing in
+        # the test suite should be going on for more than ten minutes.
+        cache_settings={
+            "uri": redis_uri,
+            "data_ttl": 600,  # 10 minutes
+            "seq_ttl": 600,  # 10 minutes
+            "socket_timeout": 600,  # 10 minutes
+            "socket_connect_timeout": 10,
+        },
+    )
+    app = build_app(
+        tree,
+        authentication=Authentication(single_user_api_key="secret"),
+    )
+
+    with Context.from_app(app) as context:
+        yield context
 
 
 @pytest.fixture